feature/task-refactor: feat: :sparkles: define the logic for the flow to publish the app in a normal pipeline flow
