+++ tasks/shared/utils/index.ts
@@ -11,4 +11,4 @@ export * from './methods'
 export * from './pullRequest'
 export * from './release'
 export * from './runCommand'
-export * from './vtexPublish'
+export * from './vtex'
