+++ tasks/shared/utils/vtex/index.ts
@@ -1,0 +1,2 @@
+export * from './publish'
+export * from './pullRequest'
