+++ tasks/vtex/pullRequest/publish/index.ts
@@ -12,14 +12,14 @@ import {
   publishAppSuccessMessage,
   setEmailAndUserGit,
   startPublishMessage,
-  vtexPublish,
+  vtexPullRequestPublish,
 } from '../../../shared'
 
 async function run() {
   try {
     // ******* Setup utilities *******
     // 1. install vtex, projex and make login in vtex with projex
-    const { forceVtexPublish } = await initialSetup()
+    const { forceVtexPublish, deploy } = await initialSetup()
     const azureConnection = await GitConnection()
     const { pullRequest } = azureConnection
     const { createdBy, sourceRefName, targetRefName } = pullRequest
@@ -50,11 +50,12 @@ async function run() {
 
     // ******* Publish *******
     // 1. make the publish
-    await vtexPublish(
+    await vtexPullRequestPublish(
       azureConnection,
       titleRelease,
       typeRelease,
-      forceVtexPublish
+      forceVtexPublish,
+      deploy
     )
 
     // 2. Complete the pull request in the main branch
